commit 8c430e5901323dc4ac1bd69a6bbf0bc356f02ae5
    3708 Fast reboot support in ixgbe
commit 187670a04e7557914566fc449b4d3af38caea282
    3666 Implement SOCK_CLOEXEC flag to socket()
commit f971a3462face662ae8ef220a18a98354d625d54
    3673 core dumping is abysmally slow
    3671 left behind enemy lines, agent LWP can go rogue
    3670 add visibility into agent LWP's spymaster
commit abc79d9dd51e98eafb6fc25b4a0b4f66bef40b00
    3679 prtconf should print out PCI database information
    3680 Want a library to allow programatic access to the pci database

	Conflicts:

	usr/src/uts/common/fs/proc/prvnops.c
	usr/src/uts/common/fs/proc/prcontrol.c
	usr/src/man/man1m/prtconf.1m
	usr/src/lib/Makefile
	usr/src/cmd/ptools/pflags/pflags.c

	Manifests:
	usr/src/pkg/manifests/system-library.mf [already present]

+++ usr/src/cmd/ptools/pflags/pflags.c
@@ -344,7 +344,7 @@ lwplook(look_arg_t *arg, const lwpstatus_t *psp, const lwpsinfo_t *pip)
 		(void) strftime(t, sizeof (t), "%F:%H.%M.%S", localtime(&time));
 
 		(void) printf("\tspymaster = pid %d, \"%s\" at %s\n",
-		    ps.pr_pid, ps.pr_psargs, t);
+		    (int)ps.pr_pid, ps.pr_psargs, t);
 	}
 
 	if (rflag) {
